Enum little::LittleError

pub enum LittleError {
    ParameterMissing(String),
    ConstantMissing(Constant),
    CallMissing(Call),
    CallError(Box<Error + Sync + Send>),
    OutputError(Error),
    BuildError(BuildError),
    StackUnderflow,
    Interupt,
}

Runtime error.

Variants

ParameterMissing

A parameter was required for an instruction, but it was not found.

ConstantMissing

A constant was required for an instruction, but it was not found.

CallMissing

A call was required for an instruction, but it was not found.

CallError

A call has returned an error.

OutputError

I/O error writing template result to output.

BuildError

Error building the template.

StackUnderflow

Attempt to pop values on empty stack.

Interupt

Instruction has caused an interupt, it is up to user to know how to handle it.